LayoutTests/webarchive/loading/missing-data.html

<script>
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.dumpAsText();
    }
    if (window.internals)
        internals.setAlwaysAllowLocalWebarchive();
    onload = function() {
        frame = document.createElement("iframe");
        frame.src = "resources/missing-data.webarchive";
        frame.onload = function() {
            if (window.testRunner)
                testRunner.notifyDone();
        }
        document.body.appendChild(frame);
    }
</script>
